Move interpret.ts to test tree
